Create Conjecture
Given a natural number num, the user generates a sequence of integers, described in the Collatz conjecture, namely if num is an even number, the program divides it in half, and if it is an odd number, the program multiplies it by 3 and adds 1. Furthermore, it repeats this operation until it gets the number 1 as a result.